# docs/api/openapi.json
{
  "components": {
...
    "schemas": {
...
      "AdjustmentRequest": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/Line"
            },
            "minItems": 1,
            "type": "array"
          },
          "postingDate": {
            "format": "date",
            "type": "string"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          },
          "reference": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "valueDate": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "lines",
          "postingDate",
          "reason",
          "reference",
          "valueDate"
        ],
        "type": "object"
      },
      "AdjustmentView": {
        "properties": {
          "entryId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CaseFileResponse": {
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckView"
            },
            "type": "array"
          },
          "customerId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "openedAt": {
            "format": "date-time",
            "type": "string"
          },
          "owners": {
            "items": {
              "$ref": "#/components/schemas/OwnerView"
            },
            "type": "array"
          },
          "policyVersion": {
            "type": "string"
          },
          "reviewTasks": {
            "items": {
              "$ref": "#/components/schemas/TaskView"
            },
            "type": "array"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CheckView": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "requestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "DecisionRequest": {
        "properties": {
          "outcome": {
            "enum": [
              "APPROVED",
              "REJECTED"
            ],
            "type": "string"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "pattern": "[^\\p{Cc}\\p{Cf}\\p{Cs}\\p{Co}\\p{Cn}]+",
            "type": "string"
          }
        },
        "required": [
          "outcome",
          "reason"
        ],
        "type": "object"
      },
...
      "Line": {
        "properties": {
          "accountId": {
            "minLength": 1,
            "type": "string"
          },
          "amount": {
            "minLength": 1,
            "type": "string"
          },
          "currency": {
            "minLength": 1,
            "pattern": "[A-Z]{3}",
            "type": "string"
          },
          "direction": {
            "enum": [
              "DEBIT",
              "CREDIT"
            ],
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "amount",
          "currency",
          "direction"
        ],
        "type": "object"
      },
      "LineView": {
        "properties": {
          "accountId": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "OwnerView": {
        "properties": {
          "controlRole": {
            "type": "string"
          },
          "declaredAt": {
            "format": "date-time",
            "type": "string"
          },
          "ownerPartyId": {
            "type": "string"
          },
          "stakeBasisPoints": {
            "format": "int32",
            "type": "integer"
          },
          "verificationPending": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "ProposalDetailView": {
        "properties": {
          "decidedAt": {
            "type": "string"
          },
          "decidedBy": {
            "type": "string"
          },
          "entryId": {
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LineView"
            },
            "type": "array"
          },
          "postingDate": {
            "type": "string"
          },
          "proposalId": {
            "type": "string"
          },
          "proposedAt": {
            "type": "string"
          },
          "proposedBy": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "valueDate": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ProposalView": {
        "properties": {
          "proposalId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ResolutionRequest": {
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "pattern": "[^\\p{Cc}\\p{Cf}\\p{Cs}\\p{Co}\\p{Cn}]+",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
      "TaskView": {
        "properties": {
          "checkId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "openedAt": {
            "format": "date-time",
            "type": "string"
          },
          "resolutionReason": {
            "type": "string"
          },
          "resolvedAt": {
            "format": "date-time",
            "type": "string"
          },
          "resolvedBy": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The public HTTP contract. Versioned in the path (ADR-0015); errors follow RFC 9457, catalogued in docs/architecture/ERROR_CONTRACT.md. Generated from the application on every build and compared against this committed copy, so a change to the contract cannot reach a client without being reviewed.",
    "title": "finapp platform API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/kyc/cases/{id}": {
      "get": {
        "operationId": "readCase",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseFileResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/v1/kyc/cases/{id}/decision": {
      "post": {
        "operationId": "decide",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/v1/kyc/cases/{id}/reviews/{taskId}/resolution": {
      "post": {
        "operationId": "resolve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/v1/ledger/adjustments": {
      "post": {
        "operationId": "propose",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
    "/v1/ledger/adjustments/{id}": {
      "delete": {
        "operationId": "reject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "get": {
        "operationId": "view_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalDetailView"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/v1/ledger/adjustments/{id}/approval": {
      "post": {
        "operationId": "approve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentView"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
...
  }
}